Remove exprimental frame pacer as not beneficial
