+++ src/platform/webos/input.rs
@@ -8,8 +8,8 @@ pub const WEBOS_BACK_KEYCODE: i32 = 2_097_155;
 
 /// Magic Remote Red button keycode. Confirmed on-device: Red arrives as a plain `KeyDown`
 /// carrying only this keycode and *no* scancode — the `SDL_SCANCODE_WEBOS_RED = 486` the SDL
-/// fork documents never shows up in the keyboard-state array, so polling it (the way Green,
-/// Yellow and Blue are read) finds nothing. Same `0x200000 + n` family as
+/// fork documents never shows up in the keyboard-state array, so polling it (the way Green and
+/// Yellow are read) finds nothing. Same `0x200000 + n` family as
 /// [`WEBOS_BACK_KEYCODE`], which behaves identically.
 pub const WEBOS_RED_KEYCODE: i32 = 2_097_169;
 
@@ -97,10 +97,9 @@ impl StickMenuNav {
 /// webOS Magic Remote scancodes — outside rust-sdl2's enum, needs raw polling.
 /// `SDL_SCANCODE_WEBOS_{RED,GREEN,YELLOW,BLUE} = 486..489` in `webosbrew/SDL-webOS`'s `SDL_scancode.h`.
 /// Red has no usable scancode here — it arrives as a bare keycode instead, see
-/// [`WEBOS_RED_KEYCODE`]. Which is also why Blue, not Red, drives the pacing toggle.
+/// [`WEBOS_RED_KEYCODE`].
 pub const WEBOS_GREEN_SCANCODE: i32 = 487;
 pub const WEBOS_YELLOW_SCANCODE: i32 = 488;
-pub const WEBOS_BLUE_SCANCODE: i32 = 489;
 
 /// webOS Home key (`SDL_SCANCODE_WEBOS_HOME = 384`). Polled to re-open the launcher once
 /// `KEYS_HOME` capture stops the OS doing it. A USB keyboard's Super key is Home-class and
